+++ cache/graph/graph.go
@@ -2,12 +2,12 @@ package graph
 
 type Graph[S comparable, T any] struct {
 	Vertices map[S]T             `json:"vertices,omitempty"`
-	Edges    map[S]map[S]float64 `json:"edges,omitempty"`
+	Edges    map[S]map[S]float32 `json:"edges,omitempty"`
 }
 
 func NewGraph[S comparable, T any]() *Graph[S, T] {
 	return &Graph[S, T]{
 		Vertices: make(map[S]T),
-		Edges:    make(map[S]map[S]float64),
+		Edges:    make(map[S]map[S]float32),
 	}
 }
